Module graph_store_postgres::block_range

Structs

BlockRange
The range of blocks for which an entity is valid. We need this struct to bind ranges into Diesel queries.

BlockRangeContainsClause
Generate the clause that checks whether block is in the block range of an entity

Constants

BLOCK_RANGE_COLUMN
The name of the column in which we store the block range

BLOCK_RANGE_CURRENT
The SQL clause we use to check that an entity version is current; that version has an unbounded block range, but checking for upper_inf(block_range) is slow and can’t use the exclusion index we have on entity tables; we therefore check if i32::MAX is in the range

BLOCK_UNVERSIONED
Most subgraph metadata entities are not versioned. For such entities, we want two things:

UNVERSIONED_RANGE

Functions

block_number
Return the block number contained in the history event. If it is None panic because that indicates that we want to perform an operation that does not record history, which should not happen with how we currently use relational schemas

clone_bound

first_block_in_range