Module graph_store_postgres::block_store

Modules

primary

Structs

BlockStore
The store that chains use to maintain their state and cache often used data from a chain. The BlockStore maintains information about all configured chains, and serves as a directory of chains, whereas the ChainStore manages chain-specific data. The BlockStore is sharded so that each chain can use, depending on configuration, a separate database. Regardless of configuration, the data for each chain is stored in its own database namespace, though for historical reasons, the code also deals with a shared table for the block and call cache for chains in the public namespace.

Enums

ChainStatus
The status of a chain: whether we can only read from the chain, or whether it is ok to ingest from it, too

Constants

FAKE_NETWORK_SHARED