Module graph_store_postgres::catalog

Modules

foreign_tables

pg_namespace

table_stats

Structs

Catalog
Information about what tables and columns we have in the database

Functions

account_like

copy_account_like

current_servers

drop_foreign_schema
Drop the schema for src if it is a foreign schema imported from another database. If the schema does not exist, or is not a foreign schema, do nothing. This crucially depends on the fact that we never mix foreign and local tables in the same schema.

get_text_columns

has_namespace

set_account_like

supports_proof_of_indexing