Module graph_store_postgres::chain_store

Re-exports

pub use data::Storage;

Modules

data
Encapuslate access to the blocks table for a chain.

public
Tables in the ‘public’ database schema that store chain-specific data

test_support
Support for tests

Structs

ChainStore

Functions

contract_call_id
The id is the hashed encoded_call + contract_address + block hash to uniquely identify the call. 256 bits of output, and therefore 128 bits of security against collisions, are needed since this could be targeted by a birthday attack.