Module graph_store_postgres::chain_store::data

Encapuslate access to the blocks table for a chain.

Modules

public

Structs

BlockHashBytea

BlockHashText

BlocksTable
The table that holds blocks when we store a chain in its own dedicated database schema

CallCacheTable

CallMetaTable

Schema

Enums

Storage
Storage for a chain. The underlying namespace (database schema) is either public or of the form chain[0-9]+.

Functions

_impl_as_expression_for_storage

_impl_from_sql_row_for_storage

_impl_queryable_by_name_for_blockhashbytea

_impl_queryable_by_name_for_blockhashtext

h256_from_bytes

Type Definitions

DynColumn

DynTable