Enum graph_store_postgres::chain_store::data::Storage

pub enum Storage {
    Shared,
    Private(Schema),
}

Storage for a chain. The underlying namespace (database schema) is either public or of the form chain[0-9]+.

Variants

Shared

Chain data is stored in shared tables

Private(Schema)

The chain has its own namespace in the database with dedicated tables

Implementations

impl Storage

const PREFIX: &'static str

const PUBLIC: &'static str

fn new(s: String) -> Result<Self, String>

pub(super) fn create(&self, conn: &PgConnection) -> Result<(), Error>

Create dedicated database tables for this chain if it uses Storage::Private. If it uses Storage::Shared, do nothing since a regular migration will already have created the ethereum_blocks table

pub(super) fn drop_storage(
&self,
conn: &PgConnection,
name: &str
) -> Result<(), StoreError>

pub(super) fn upsert_block(
&self,
conn: &PgConnection,
chain: &str,
block: EthereumBlock
) -> Result<(), StoreError>

Insert a block. If the table already contains a block with the same hash, then overwrite that block since it may be adding transaction receipts.

pub(super) fn upsert_light_block(
&self,
conn: &PgConnection,
chain: &str,
block: LightEthereumBlock
) -> Result<(), Error>

Insert a light block. On conflict do nothing, since we do not want to erase transaction receipts that might already be there

pub(super) fn blocks(
&self,
conn: &PgConnection,
chain: &str,
hashes: Vec<H256>
) -> Result<Vec<LightEthereumBlock>, Error>

pub(super) fn block_hashes_by_block_number(
&self,
conn: &PgConnection,
chain: &str,
number: BlockNumber
) -> Result<Vec<H256>, Error>

pub(super) fn confirm_block_hash(
&self,
conn: &PgConnection,
chain: &str,
number: BlockNumber,
hash: &H256
) -> Result<usize, Error>

pub(super) fn block_number(
&self,
conn: &PgConnection,
hash: H256
) -> Result<Option<BlockNumber>, StoreError>

pub(super) fn missing_parent(
&self,
conn: &PgConnection,
chain: &str,
first_block: i64,
hash: H256,
genesis: H256
) -> Result<Option<H256>, Error>

Find the first block that is missing from the database needed to complete the chain from block hash to the block with number first_block.

pub(super) fn chain_head_candidate(
&self,
conn: &PgConnection,
chain: &str
) -> Result<Option<BlockPtr>, Error>

Return the best candidate for the new chain head if there is a block with a higher block number than the current chain head. The returned value if the hash and number of the candidate and the genesis block hash for the chain

pub(super) fn ancestor_block(
&self,
conn: &PgConnection,
block_ptr: BlockPtr,
offset: BlockNumber
) -> Result<Option<EthereumBlock>, Error>

pub(super) fn delete_blocks_before(
&self,
conn: &PgConnection,
chain: &str,
block: i64
) -> Result<usize, Error>

pub(super) fn get_call_and_access(
&self,
conn: &PgConnection,
id: &[u8]
) -> Result<Option<(Vec<u8>, bool)>, Error>

pub(super) fn update_accessed_at(
&self,
conn: &PgConnection,
contract_address: &[u8]
) -> Result<(), Error>

pub(super) fn set_call(
&self,
conn: &PgConnection,
id: &[u8],
contract_address: &[u8],
block_number: i32,
return_value: &[u8]
) -> Result<(), Error>

pub(super) fn set_chain(
&self,
conn: &PgConnection,
chain_name: &str,
genesis_hash: &str,
chain: Chain
)

pub(crate) fn find_transaction_receipts_in_block(
&self,
conn: &PgConnection,
schema_name: &str,
block_hash: &H256
) -> Result<Vec<LightTransactionReceipt>>

Delegates to [transaction_receipt::find_transaction_receipts_in_block].