Module graph_store_postgres::chain_store::test_support

Support for tests

Structs

FakeBlock
The parts of an Ethereum block that are interesting for these tests: the block number, hash, and the hash of the parent block

Constants

NO_PARENT

Traits

SettableChainStore
Store the given chain as the blocks for the network set the network’s genesis block to genesis_hash, and head block to null

Type Definitions

Chain