Module graph_store_postgres::command_support::catalog

Modules

active_copies

block_store

copy_state

copy_table_state

deployment_schemas

ens_names

subgraph

subgraph_deployment_assignment

subgraph_version

Structs

Connection
A wrapper for a database connection that provides access to functionality that works only on the primary database

Site
Details about a deployment and the shard in which it is stored. We need the database namespace for the deployment as that information is only stored in the primary database.

Functions

account_like

set_account_like