Module graph_store_postgres::command_support::catalog::copy_table_state::dsl

Re-exports all of the columns of this table, as well as the table struct renamed to the module name. This is meant to be glob imported for functions which only deal with one table.

Structs

batch_size

copy_table_state
The actual table struct

dst

duration_ms

entity_type

finished_at

id

next_vid

started_at

target_vid