Struct graph_store_postgres::command_support::catalog::Connection

pub struct Connection<'a>(MaybeOwned<'a, PooledConnection<ConnectionManager<PgConnection>>>);

A wrapper for a database connection that provides access to functionality that works only on the primary database

Implementations

impl<'a> Connection<'a>

pub fn new(
conn: impl Into<MaybeOwned<'a, PooledConnection<ConnectionManager<PgConnection>>>>
) -> Self

pub(crate) fn transaction<T, E, F>(&self, f: F) -> Result<T, E> where
F: FnOnce() -> Result<T, E>,
E: From<Error>,

pub fn current_deployment_for_subgraph(
&self,
name: SubgraphName
) -> Result<DeploymentHash, StoreError>

fn cancel_copies(&self, ids: Vec<DeploymentId>) -> Result<(), StoreError>

Signal any copy process that might be copying into one of these deployments that it should stop. Copying is cancelled whenever we remove the assignment for a deployment

fn remove_unused_assignments(&self) -> Result<Vec<EntityChange>, StoreError>

Delete all assignments for deployments that are neither the current nor the pending version of a subgraph and return the deployment id’s

pub fn promote_deployment(
&self,
id: &DeploymentHash
) -> Result<Vec<EntityChange>, StoreError>

Promote the deployment id to the current version everywhere where it was the pending version so far, and remove any assignments that are not needed any longer as a result. Return the changes that were made to assignments in the process

pub fn create_subgraph(&self, name: &SubgraphName) -> Result<String, StoreError>

Create a new subgraph with the given name. If one already exists, use the existing one. Return the id of the newly created or existing subgraph

pub fn create_subgraph_version<F>(
&self,
name: SubgraphName,
site: &Site,
node_id: NodeId,
mode: SubgraphVersionSwitchingMode,
exists_and_synced: F
) -> Result<Vec<EntityChange>, StoreError> where
F: FnOnce(&DeploymentHash) -> Result<bool, StoreError>,

pub fn remove_subgraph(
&self,
name: SubgraphName
) -> Result<Vec<EntityChange>, StoreError>

pub fn subgraph_exists(&self, name: &SubgraphName) -> Result<bool, StoreError>

pub fn reassign_subgraph(
&self,
site: &Site,
node: &NodeId
) -> Result<Vec<EntityChange>, StoreError>

pub fn assign_subgraph(
&self,
site: &Site,
node: &NodeId
) -> Result<Vec<EntityChange>, StoreError>

pub fn unassign_subgraph(
&self,
site: &Site
) -> Result<Vec<EntityChange>, StoreError>

fn create_site(
&self,
shard: Shard,
deployment: DeploymentHash,
network: String,
active: bool
) -> Result<Site, StoreError>

Create a new site and possibly set it to the active site. This function only performs the basic operations for creation, and the caller must check that other conditions (like whether there already is an active site for the deployment) are met

pub fn allocate_site(
&self,
shard: Shard,
subgraph: &DeploymentHash,
network: String
) -> Result<Site, StoreError>

pub fn copy_site(&self, src: &Site, shard: Shard) -> Result<Site, StoreError>

Create a copy of the site src in the shard shard, but mark it as not active. If there already is a site in shard, return that instead.

pub(crate) fn activate(
&self,
deployment: &DeploymentLocator
) -> Result<(), StoreError>

pub fn drop_site(&self, site: &Site) -> Result<(), StoreError>

Remove all subgraph versions and the entry in deployment_schemas for subgraph id in a transaction

pub fn find_active_site(
&self,
subgraph: &DeploymentHash
) -> Result<Option<Site>, StoreError>

pub fn find_site_in_shard(
&self,
subgraph: &DeploymentHash,
shard: &Shard
) -> Result<Option<Site>, StoreError>

pub fn find_site_by_ref(
&self,
id: DeploymentId
) -> Result<Option<Site>, StoreError>

pub fn find_site_by_name(&self, name: &str) -> Result<Option<Site>, StoreError>

pub fn find_sites_for_network(
&self,
network: &str
) -> Result<Vec<Site>, StoreError>

pub fn find_sites(
&self,
ids: Vec<String>,
only_active: bool
) -> Result<Vec<Site>, StoreError>

Find sites by their subgraph deployment hashes. If ids is empty, return all sites

pub fn find_sites_by_id(
&self,
ids: Vec<DeploymentId>
) -> Result<Vec<Site>, StoreError>

Find sites by their subgraph deployment ids. If ids is empty, return all sites

pub fn sites(&self) -> Result<Vec<Site>, StoreError>

pub fn send_store_event(&self, event: &StoreEvent) -> Result<(), StoreError>

pub fn least_assigned_node(
&self,
nodes: &Vec<NodeId>
) -> Result<Option<NodeId>, StoreError>

Return the name of the node that has the fewest assignments out of the given nodes. If nodes is empty, return None

pub fn assigned_node(&self, site: &Site) -> Result<Option<NodeId>, StoreError>

pub fn assignments(&self, node: &NodeId) -> Result<Vec<Site>, StoreError>

pub fn fill_assignments(
&self,
infos: Vec<Info>
) -> Result<Vec<Info>, StoreError>

pub(crate) fn deployments_for_subgraph(
&self,
name: String
) -> Result<Vec<Site>, StoreError>

pub fn subgraph_version(
&self,
name: String,
use_current: bool
) -> Result<Option<Site>, StoreError>

pub fn versions_for_subgraph(
&self,
name: &str
) -> Result<(Option<String>, Option<String>), StoreError>

pub fn deployment_for_version(
&self,
name: &str
) -> Result<Option<String>, StoreError>

pub fn version_info(
&self,
version: &str
) -> Result<Option<(String, String)>, StoreError>

pub fn versions_for_subgraph_id(
&self,
subgraph_id: &str
) -> Result<(Option<String>, Option<String>), StoreError>

pub fn detect_unused_deployments(&self) -> Result<Vec<Site>, StoreError>

Find all deployments that are not in use and add them to the unused_deployments table. Only values that are available in the primary will be filled in unused_deployments

pub fn update_unused_deployments(
&self,
details: &Vec<DeploymentDetail>
) -> Result<(), StoreError>

Add details from the deployment shard to unused deployments

pub fn list_unused_deployments(
&self,
filter: Filter
) -> Result<Vec<UnusedDeployment>, StoreError>

pub fn subgraphs_using_deployment(
&self,
site: &Site
) -> Result<Vec<String>, StoreError>

pub fn find_ens_name(&self, hash: &str) -> Result<Option<String>, StoreError>

pub fn record_active_copy(
&self,
src: &Site,
dst: &Site
) -> Result<(), StoreError>

pub fn copy_finished(&self, dst: &Site) -> Result<(), StoreError>