Struct graph_store_postgres::connection_pool::ConnectionPool

pub struct ConnectionPool {
    logger: Logger,
    shard: Shard,
    pool: Pool<ConnectionManager<PgConnection>>,
    fdw_pool: Option<Pool<ConnectionManager<PgConnection>>>,
    limiter: Arc<Semaphore>,
    postgres_url: String,
    pub(crate) wait_stats: PoolWaitStats,
    query_semaphore: Arc<Semaphore>,
    semaphore_wait_stats: Arc<RwLock<MovingStats>>,
    semaphore_wait_gauge: Box<Gauge>,
}

Fields
logger: Loggershard: Shardpool: Pool<ConnectionManager<PgConnection>>fdw_pool: Option<Pool<ConnectionManager<PgConnection>>>limiter: Arc<Semaphore>postgres_url: Stringwait_stats: PoolWaitStatsquery_semaphore: Arc<Semaphore>semaphore_wait_stats: Arc<RwLock<MovingStats>>semaphore_wait_gauge: Box<Gauge>
Implementations

impl ConnectionPool

pub fn create(
shard_name: &str,
pool_name: &str,
postgres_url: String,
pool_size: u32,
fdw_pool_size: Option<u32>,
logger: &Logger,
registry: Arc<dyn MetricsRegistry>
) -> ConnectionPool

pub(crate) async fn with_conn<T: Send + 'static>(
&self,
f: impl 'static + Send + FnOnce(&PooledConnection<ConnectionManager<PgConnection>>, &CancelHandle) -> Result<T, CancelableError<StoreError>>
) -> Result<T, StoreError>

Execute a closure with a connection to the database.

API

The API of using a closure to bound the usage of the connection serves several purposes:

Moves blocking database access out of the Future::poll. Within Future::poll (which includes all async methods) it is illegal to perform a blocking operation. This includes all accesses to the database, acquiring of locks, etc. Calling a blocking operation can cause problems with Future combinators (including but not limited to select, timeout, and FuturesUnordered) and problems with executors/runtimes. This method moves the database work onto another thread in a way which does not block Future::poll.

Limit the total number of connections. Because the supplied closure takes a reference, we know the scope of the usage of all entity connections and can limit their use in a non-blocking way.

Cancellation

The normal pattern for futures in Rust is drop to cancel. Once we spawn the database work in a thread though, this expectation no longer holds because the spawned task is the independent of this future. So, this method provides a cancel token which indicates that the Future has been dropped. This isn’t quite as good as drop on cancel, because a drop on cancel can do things like cancel http requests that are in flight, but checking for cancel periodically is a significant improvement.

The implementation of the supplied closure should check for cancel between every operation that is potentially blocking. This includes any method which may interact with the database. The check can be conveniently written as token.check_cancel()?;. It is low overhead to check for cancel, so when in doubt it is better to have too many checks than too few.

Panics:

This task will panic if the supplied closure panics

This task will panic if the supplied closure returns Err(Cancelled) when the supplied cancel token is not cancelled.

pub fn get_with_timeout_warning(
&self,
logger: &Logger
) -> Result<PooledConnection<ConnectionManager<PgConnection>>, Error>

pub fn get_fdw<F>(
&self,
logger: &Logger,
timeout: F
) -> Result<PooledConnection<ConnectionManager<PgConnection>>, Error> where
F: FnMut() -> bool,

Get a connection from the pool for foreign data wrapper access; since that pool can be very contended, periodically log that we are still waiting for a connection

The timeout is called every time we time out waiting for a connection. If timeout returns true, get_fdw returns with that error, otherwise we try again to get a connection.

pub fn connection_detail(&self) -> Result<ForeignServer, StoreError>

pub async fn setup(&self, servers: Arc<Vec<ForeignServer>>)

Setup the database for this pool. This includes configuring foreign data wrappers for cross-shard communication, and running any pending schema migrations for this database.

Panics

If any errors happen during the migration, the process panics

pub(crate) async fn query_permit(&self) -> OwnedSemaphorePermit

fn configure_fdw(&self, servers: &Vec<ForeignServer>) -> Result<(), StoreError>

fn map_primary(&self) -> Result<(), StoreError>

Map key tables from the primary into our local schema. If we are the primary, set them up as views.

We recreate this mapping on every server start so that migrations that change one of the mapped tables actually show up in the imported tables

fn map_metadata(&self, servers: &Vec<ForeignServer>) -> Result<(), StoreError>