Module graph_store_postgres::copy

This module if repsonsible for copying the data of an existing subgraph src to a new subgraph dst. The copy is done in batches so that copying does not cause long-running transactions since they lead to table bloat in the rest of the system.

The TARGET_DURATION defines how long any individual copy operation should take at most, and the code adapts the number of entities that are copied in each batch to be close to that time span.

The progress of copy operations is recorded in the tables subgraphs.copy_state and subgraphs.copy_table_state so that a copy operation can resume after an interruption, for example, because graph-node was restarted while the copy was running.

Modules

active_copies

copy_state

copy_table_state

Structs

Connection
A helper for copying subgraphs

CopyProgress

CopyState

TableState

Enums

Status

Constants

INITIAL_BATCH_SIZE

LOG_INTERVAL

TARGET_DURATION