Function graph_store_postgres::deployment::graft_pending

pub fn graft_pending(
    conn: &PgConnection, 
    id: &DeploymentHash
) -> Result<Option<(DeploymentHash, BlockPtr)>, StoreError>

Look up the graft point for the given subgraph in the database and return it. Returns None if the deployment does not have a graft or if the subgraph has already progress past the graft point, indicating that the data copying for grafting has been performed