Struct graph_store_postgres::deployment_store::DeploymentStore

pub struct DeploymentStore(Arc<StoreInner>);

Storage of the data for individual deployments. Each DeploymentStore corresponds to one of the database shards that SubgraphStore manages.

Implementations

impl DeploymentStore

pub fn new(
logger: &Logger,
pool: ConnectionPool,
read_only_pools: Vec<ConnectionPool>,
pool_weights: Vec<usize>
) -> Self

pub(crate) fn create_deployment(
&self,
schema: &Schema,
deployment: SubgraphDeploymentEntity,
site: Arc<Site>,
graft_base: Option<Arc<Layout>>,
replace: bool
) -> Result<(), StoreError>

pub(crate) fn load_deployment(
&self,
site: &Site
) -> Result<SubgraphDeploymentEntity, StoreError>

pub(crate) fn drop_deployment(&self, site: &Site) -> Result<(), StoreError>

pub(crate) fn execute_query<T: FromEntityData>(
&self,
conn: &PgConnection,
site: Arc<Site>,
query: EntityQuery
) -> Result<Vec<T>, QueryExecutionError>

fn check_interface_entity_uniqueness(
&self,
conn: &PgConnection,
layout: &Layout,
key: &EntityKey
) -> Result<(), StoreError>

fn apply_entity_modifications(
&self,
conn: &PgConnection,
layout: &Layout,
mods: Vec<EntityModification>,
ptr: &BlockPtr,
stopwatch: StopwatchMetrics
) -> Result<i32, StoreError>

fn insert_entities(
&self,
entity_type: &EntityType,
data: &mut [(EntityKey, Entity)],
conn: &PgConnection,
layout: &Layout,
ptr: &BlockPtr,
stopwatch: &StopwatchMetrics
) -> Result<usize, StoreError>

fn overwrite_entities(
&self,
entity_type: &EntityType,
data: &mut [(EntityKey, Entity)],
conn: &PgConnection,
layout: &Layout,
ptr: &BlockPtr,
stopwatch: &StopwatchMetrics
) -> Result<usize, StoreError>

fn remove_entities(
&self,
entity_type: &EntityType,
entity_keys: &[String],
conn: &PgConnection,
layout: &Layout,
ptr: &BlockPtr,
stopwatch: &StopwatchMetrics
) -> Result<usize, StoreError>

pub(crate) async fn with_conn<T: Send + 'static>(
&self,
f: impl 'static + Send + FnOnce(&PooledConnection<ConnectionManager<PgConnection>>, &CancelHandle) -> Result<T, CancelableError<StoreError>>
) -> Result<T, StoreError>

Execute a closure with a connection to the database.

API

The API of using a closure to bound the usage of the connection serves several purposes:

Moves blocking database access out of the Future::poll. Within Future::poll (which includes all async methods) it is illegal to perform a blocking operation. This includes all accesses to the database, acquiring of locks, etc. Calling a blocking operation can cause problems with Future combinators (including but not limited to select, timeout, and FuturesUnordered) and problems with executors/runtimes. This method moves the database work onto another thread in a way which does not block Future::poll.

Limit the total number of connections. Because the supplied closure takes a reference, we know the scope of the usage of all entity connections and can limit their use in a non-blocking way.

Cancellation

The normal pattern for futures in Rust is drop to cancel. Once we spawn the database work in a thread though, this expectation no longer holds because the spawned task is the independent of this future. So, this method provides a cancel token which indicates that the Future has been dropped. This isn’t quite as good as drop on cancel, because a drop on cancel can do things like cancel http requests that are in flight, but checking for cancel periodically is a significant improvement.

The implementation of the supplied closure should check for cancel between every operation that is potentially blocking. This includes any method which may interact with the database. The check can be conveniently written as token.check_cancel()?;. It is low overhead to check for cancel, so when in doubt it is better to have too many checks than too few.

Panics:

This task will panic if the supplied closure panics

This task will panic if the supplied closure returns Err(Cancelled) when the supplied cancel token is not cancelled.

fn get_conn(
&self
) -> Result<PooledConnection<ConnectionManager<PgConnection>>, Error>

Deprecated. Use with_conn instead.

fn read_only_conn(
&self,
idx: usize
) -> Result<PooledConnection<ConnectionManager<PgConnection>>, Error>

Panics if idx is not a valid index for a read only pool.

pub(crate) fn get_replica_conn(
&self,
replica: ReplicaId
) -> Result<PooledConnection<ConnectionManager<PgConnection>>, Error>

pub(crate) async fn query_permit(
&self,
replica: ReplicaId
) -> OwnedSemaphorePermit

pub(crate) fn wait_stats(&self, replica: ReplicaId) -> &PoolWaitStats

pub(crate) fn layout(
&self,
conn: &PgConnection,
site: Arc<Site>
) -> Result<Arc<Layout>, StoreError>

Return the layout for a deployment. Since constructing a Layout object takes a bit of computation, we cache layout objects that do not have a pending migration in the Store, i.e., for the lifetime of the Store. Layout objects with a pending migration can not be cached for longer than a transaction since they might change without us knowing

pub(crate) fn find_layout(
&self,
site: Arc<Site>
) -> Result<Arc<Layout>, StoreError>

Return the layout for a deployment. This might use a database connection for the lookup and should only be called if the caller does not have a connection currently. If it does, use layout

fn subgraph_info_with_conn(
&self,
conn: &PgConnection,
site: &Site
) -> Result<SubgraphInfo, StoreError>

pub(crate) fn subgraph_info(
&self,
site: &Site
) -> Result<SubgraphInfo, StoreError>

fn block_ptr_with_conn(
subgraph_id: &DeploymentHash,
conn: &PgConnection
) -> Result<Option<BlockPtr>, Error>

pub(crate) fn deployment_details(
&self,
ids: Vec<String>
) -> Result<Vec<DeploymentDetail>, StoreError>

pub(crate) fn deployment_statuses(
&self,
sites: &Vec<Arc<Site>>
) -> Result<Vec<Info>, StoreError>

pub(crate) fn deployment_exists_and_synced(
&self,
id: &DeploymentHash
) -> Result<bool, StoreError>

pub(crate) fn deployment_synced(
&self,
id: &DeploymentHash
) -> Result<(), StoreError>

pub(crate) fn drop_deployment_schema(
&self,
namespace: &Namespace
) -> Result<(), StoreError>

pub(crate) fn drop_all_metadata(&self) -> Result<(), StoreError>

pub(crate) async fn vacuum(&self) -> Result<(), StoreError>

impl DeploymentStore

Methods that back the trait graph::components::Store, but have small variations in their signatures

pub(crate) fn block_ptr(&self, site: &Site) -> Result<Option<BlockPtr>, Error>

pub(crate) fn supports_proof_of_indexing<'a>(
self: Arc<Self>,
site: Arc<Site>
) -> DynTryFuture<'a, bool>

pub(crate) fn get_proof_of_indexing<'a>(
self: Arc<Self>,
site: Arc<Site>,
indexer: &'a Option<Address>,
block: BlockPtr
) -> DynTryFuture<'a, Option<[u8; 32]>>

pub(crate) fn get(
&self,
site: Arc<Site>,
key: EntityKey
) -> Result<Option<Entity>, QueryExecutionError>

pub(crate) fn get_many(
&self,
site: Arc<Site>,
ids_for_type: BTreeMap<&EntityType, Vec<&str>>
) -> Result<BTreeMap<EntityType, Vec<Entity>>, StoreError>

pub(crate) fn find(
&self,
site: Arc<Site>,
query: EntityQuery
) -> Result<Vec<Entity>, QueryExecutionError>

pub(crate) fn transact_block_operations(
&self,
site: Arc<Site>,
block_ptr_to: BlockPtr,
mods: Vec<EntityModification>,
stopwatch: StopwatchMetrics,
data_sources: Vec<StoredDynamicDataSource>,
deterministic_errors: Vec<SubgraphError>
) -> Result<StoreEvent, StoreError>

fn rewind_with_conn(
&self,
conn: &PgConnection,
site: Arc<Site>,
block_ptr_to: BlockPtr
) -> Result<StoreEvent, StoreError>

pub(crate) fn rewind(
&self,
site: Arc<Site>,
block_ptr_to: BlockPtr
) -> Result<StoreEvent, StoreError>

pub(crate) fn revert_block_operations(
&self,
site: Arc<Site>,
block_ptr_to: BlockPtr
) -> Result<StoreEvent, StoreError>

pub(crate) async fn deployment_state_from_id(
&self,
id: DeploymentHash
) -> Result<DeploymentState, StoreError>

pub(crate) async fn fail_subgraph(
&self,
id: DeploymentHash,
error: SubgraphError
) -> Result<(), StoreError>

pub(crate) fn replica_for_query(
&self,
for_subscription: bool
) -> Result<ReplicaId, StoreError>

pub(crate) async fn load_dynamic_data_sources(
&self,
id: DeploymentHash
) -> Result<Vec<StoredDynamicDataSource>, StoreError>

pub(crate) async fn exists_and_synced(
&self,
id: DeploymentHash
) -> Result<bool, StoreError>

pub(crate) fn graft_pending(
&self,
id: &DeploymentHash
) -> Result<Option<(DeploymentHash, BlockPtr)>, StoreError>

pub(crate) fn start_subgraph(
&self,
logger: &Logger,
site: Arc<Site>,
graft_src: Option<(Arc<Layout>, BlockPtr)>
) -> Result<(), StoreError>

Bring the subgraph into a state where we can start or resume indexing.

If graft_src is Some(..), copy data from that subgraph. It should only be Some(..) if we know we still need to copy data. The code is idempotent so that a copy process that has been interrupted can be resumed seamlessly, but the code sets the block pointer back to the graph point, so that calling this needlessly with Some(..) will remove any progress that might have been made since the last time the deployment was started.

pub(crate) fn unfail(&self, site: Arc<Site>) -> Result<(), StoreError>

pub fn error_count(&self, id: &DeploymentHash) -> Result<usize, StoreError>