Module graph_store_postgres::primary

Utilities for dealing with subgraph metadata that resides in the primary shard. Anything in this module can only be used with a database connection for the primary shard.

Re-exports

pub use self::db_enum_impl_DeploymentSchemaVersion::DeploymentSchemaVersionMapping;

Modules

active_copies

db_enum_impl_DeploymentSchemaVersion

deployment_schemas

ens_names

subgraph

subgraph_deployment_assignment

subgraph_version

unused_deployments
A table to track deployments that are no longer used. Once an unused deployment has been removed, the entry in this table is the only trace in the system that it ever existed

Structs

Connection
A wrapper for a database connection that provides access to functionality that works only on the primary database

DeploymentId
A marker that an i32 references a deployment. Values of this type hold the primary key from the deployment_schemas table

EVENT_TAP

EVENT_TAP_ENABLED
Tests set this to true so that send_store_event will store a copy of each event sent in EVENT_TAP

Namespace
A namespace (schema) in the database

Schema
Information about the database schema that stores the entities for a subgraph.

Site
Details about a deployment and the shard in which it is stored. We need the database namespace for the deployment as that information is only stored in the primary database.

UnusedDeployment

Enums

DeploymentSchemaVersion
We used to support different layout schemes. The old ‘Split’ scheme which used JSONB layout has been removed, and we will only deal with relational layout. Trying to do anything with a ‘Split’ subgraph will result in an error.

Functions

_impl_as_expression_for_deploymentid

_impl_as_expression_for_namespace

_impl_from_sql_row_for_deploymentid

_impl_from_sql_row_for_namespace

_impl_queryable_by_name_for_schema

_impl_queryable_by_name_for_unuseddeployment

_impl_queryable_for_schema

_impl_queryable_for_unuseddeployment

make_dummy_site
This is only used for tests to allow them to create a Site that does not originate in the database