Module graph_store_postgres::relational

Support for storing the entities of a subgraph in a relational schema, i.e., one where each entity type gets its own table, and the table structure follows the structure of the GraphQL type, using the native SQL types that most appropriately map to the corresponding GraphQL types

The pivotal struct in this module is the Layout which handles all the information about mapping a GraphQL schema to database tables

Re-exports

pub use crate::catalog::Catalog;

Structs

ACCOUNT_TABLES
Deprecated; use ‘graphman stats account-like’ instead. A list of fully qualified table names that contain entities that are like accounts in that they have a relatively small number of entities, with a large number of change for each entity. It is useful to treat such tables special in queries by changing the clause that selects for a specific block range in a way that makes the BRIN index on block_range usable

CacheEntry

Column

EnumType
A user-defined enum

Layout

LayoutCache
Cache layouts for some time and refresh them when they expire. Refreshing happens one at a time, and the cache makes sure we minimize blocking while a refresh happens, favoring using an expired layout over a refreshed one.

STATEMENT_TIMEOUT
GRAPH_SQL_STATEMENT_TIMEOUT is the timeout for queries in seconds. If it is not set, no statement timeout will be enforced. The statement timeout is local, i.e., can only be used within a transaction and will be cleared at the end of the transaction

SqlName
A string we use as a SQL name for a table or column. The important thing is that SQL names are snake cased. Using this type makes it easier to spot cases where we use a GraphQL name like ‘bigThing’ when we should really use the SQL version ‘big_thing’

Table

Enums

ColumnType
This is almost the same as graph::data::store::ValueType, but without ID and List; with this type, we only care about scalar types that directly correspond to Postgres scalar types

IdType
The SQL type to use for GraphQL ID properties. We support strings and byte arrays

Constants

DELETE_OPERATION_CHUNK_SIZE

POSTGRES_MAX_PARAMETERS

PRIMARY_KEY_COLUMN
The name for the primary key column of a table; hardcoded for now

STRING_PREFIX_SIZE
The size of string prefixes that we index. This is chosen so that we will index strings that people will do string comparisons like = or != on; if text longer than this is stored in a String attribute it is highly unlikely that they will be used for exact string operations. This also makes sure that we do not put strings into a BTree index that’s bigger than Postgres’ limit on such strings which is about 2k

VID_COLUMN
We give every version of every entity in our tables, i.e., every row, a synthetic primary key. This is the name of the column we use.

Functions

is_object_type

named_type
Return the enclosed named type for a field type, i.e., the type after stripping List and NonNull.

Type Definitions

EnumMap

IdTypeMap