Trait graph_store_postgres::relational_queries::ForeignKeyClauses

trait ForeignKeyClauses {
    fn column_type(&self) -> &ColumnType;

    fn name(&self) -> &str;

    fn bind_id(&self, id: &str, out: &mut AstPass<'_, Pg>) -> QueryResult<()> { ... }

    fn bind_ids<S>(
        &self, 
        ids: &[S], 
        out: &mut AstPass<'_, Pg>
    ) -> QueryResult<()>
    where
        S: AsRef<str> + ToSql<Text, Pg>,
    { ... }

    fn eq(&self, id: &str, out: &mut AstPass<'_, Pg>) -> QueryResult<()> { ... }

    fn is_in<S>(&self, ids: &[S], out: &mut AstPass<'_, Pg>) -> QueryResult<()>
    where
        S: AsRef<str> + ToSql<Text, Pg>,
    { ... }

    fn push_matrix(
        &self, 
        matrix: &Vec<Vec<Option<SafeString>>>, 
        out: &mut AstPass<'_, Pg>
    ) -> QueryResult<()> { ... }
}

Conveniences for handling foreign keys depending on whether we are using IdType::Bytes or IdType::String as the primary key

This trait adds some capabilities to Column that are very specific to how we generate SQL queries. Using a method like bind_ids from this trait on a given column means “send these values to the database in a form that can later be used for comparisons with that column”

Required methods

fn column_type(&self) -> &ColumnType

The type of the column

fn name(&self) -> &str

The name of the column

Loading content...
Provided methods

fn bind_id(&self, id: &str, out: &mut AstPass<'_, Pg>) -> QueryResult<()>

Add id as a bind variable to out, using the right SQL type

fn bind_ids<S>(&self, ids: &[S], out: &mut AstPass<'_, Pg>) -> QueryResult<()> where
S: AsRef<str> + ToSql<Text, Pg>,

Add ids as a bind variable to out, using the right SQL type

fn eq(&self, id: &str, out: &mut AstPass<'_, Pg>) -> QueryResult<()>

Generate a clause {name()} = $id using the right types to bind $id into out

fn is_in<S>(&self, ids: &[S], out: &mut AstPass<'_, Pg>) -> QueryResult<()> where
S: AsRef<str> + ToSql<Text, Pg>,

Generate a clause exists (select 1 from unnest($ids) as p(g$id) where id = p.g$id) using the right types to bind $ids into out

fn push_matrix(
&self,
matrix: &Vec<Vec<Option<SafeString>>>,
out: &mut AstPass<'_, Pg>
) -> QueryResult<()>

Generate an array of arrays as literal SQL. The ids must form a valid matrix, i.e. the same numbe of entries in each row. This can be achieved by padding them with None values. Diesel does not support arrays of arrays as bind variables, nor arrays containing nulls, so we have to manually serialize the ids as literal SQL.

Loading content...