Struct graph_store_postgres::SubgraphStore

pub struct SubgraphStore {
    inner: Arc<SubgraphStoreInner>,
}

Multiplex store operations on subgraphs and deployments between a primary and any number of additional storage shards. The primary contains information about named subgraphs, and how the underlying deployments are spread across shards, while the actual deployment data and metadata is stored in the shards. Depending on the configuration, the database for the primary and for the shards can be the same database, in which case they are all backed by one connection pool, or separate databases in the same Postgres cluster, or entirely separate clusters. Details of how to configure shards can be found in this document

The primary uses the following database tables:

public.deployment_schemas: immutable data about deployments, including the shard that stores the deployment data and metadata, the namespace in the shard that contains the deployment data, and the network/chain that the deployment is indexing

subgraphs.subgraph and subgraphs.subgraph_version: information about named subgraphs and how they map to deployments

subgraphs.subgraph_deployment_assignment: which index node is indexing what deployment

The primary is also the database that is used to send and receive notifications through Postgres’ LISTEN/NOTIFY mechanism. That is used to send notifications about new blocks that a block ingestor has discovered, and to send StoreEvents, which are used to broadcast changes in deployment assignments and changes in subgraph data to trigger updates on GraphQL subscriptions.

For each deployment, the corresponding shard contains a namespace for the deployment data; the schema in that namespace is generated from the deployment’s GraphQL schema by the crate::relational::Layout, which is also responsible for modifying and querying subgraph data. Deployment metadata is stored in tables in the subgraphs namespace in the same shard as the deployment data. The most important of these tables are

subgraphs.subgraph_deployment: the main table for deployment metadata; most importantly, it stores the pointer to the current subgraph head, i.e., the block up to which the subgraph has indexed the chain, together with other things like whether the subgraph has synced, whether it has failed and whether it encountered any errors

subgraphs.subgraph_manifest: immutable information derived from the YAML manifest for the deployment

subgraphs.dynamic_ethereum_contract_data_source: the data sources that the subgraph has created from templates in the manifest.

subgraphs.subgraph_error: details about errors that the deployment has encountered

The SubgraphStore mostly orchestrates access to the primary and the shards. The actual work is done by code in the primary module for queries against the primary store, and by the DeploymentStore for access to deployment data and metadata.

Fields
inner: Arc<SubgraphStoreInner>
Implementations

impl SubgraphStore

pub fn new(
logger: &Logger,
stores: Vec<(Shard, ConnectionPool, Vec<ConnectionPool>, Vec<usize>)>,
placer: Arc<dyn DeploymentPlacer + Send + Sync + 'static>
) -> Self

Create a new store for subgraphs that distributes deployments across multiple databases

stores is a list of the shards. The tuple contains the shard name, the main connection pool for the database, a list of read-only connections for the same database, and a list of weights determining how often to use the main pool and the read replicas for queries. The list of weights must be one longer than the list of read replicas, and weights[0] is used for the main pool.

All write operations for a shard are performed against the main pool. One of the shards must be named primary

The placer determines where create_subgraph_deployment puts a new deployment

pub(crate) fn get_proof_of_indexing<'a>(
&self,
id: &'a DeploymentHash,
indexer: &'a Option<Address>,
block: BlockPtr
) -> DynTryFuture<'a, Option<[u8; 32]>>