Struct graph_store_postgres::subgraph_store::SubgraphStoreInner

pub struct SubgraphStoreInner {
    logger: Logger,
    primary: ConnectionPool,
    stores: HashMap<Shard, Arc<DeploymentStore>>,
    sites: TimedCache<DeploymentHash, Site>,
    placer: Arc<dyn DeploymentPlacer + Send + Sync + 'static>,
}

Fields
logger: Loggerprimary: ConnectionPoolstores: HashMap<Shard, Arc<DeploymentStore>>sites: TimedCache<DeploymentHash, Site>

Cache for the mapping from deployment id to shard/namespace/id. Only active sites are cached here to ensure we have a unique mapping from SubgraphDeploymentId to Site. The cache keeps entry only for SITES_CACHE_TTL so that changes, in particular, activation of a different deployment for the same hash propagate across different graph-node processes over time.

placer: Arc<dyn DeploymentPlacer + Send + Sync + 'static>
Implementations

impl SubgraphStoreInner

pub fn new(
logger: &Logger,
stores: Vec<(Shard, ConnectionPool, Vec<ConnectionPool>, Vec<usize>)>,
placer: Arc<dyn DeploymentPlacer + Send + Sync + 'static>
) -> Self

Create a new store for subgraphs that distributes deployments across multiple databases

stores is a list of the shards. The tuple contains the shard name, the main connection pool for the database, a list of read-only connections for the same database, and a list of weights determining how often to use the main pool and the read replicas for queries. The list of weights must be one longer than the list of read replicas, and weights[0] is used for the main pool.

All write operations for a shard are performed against the main pool. One of the shards must be named primary

The placer determines where create_subgraph_deployment puts a new deployment

pub(crate) fn clear_caches(&self)

pub fn shard(&self, deployment: &DeploymentLocator) -> Result<Shard, StoreError>

fn cache_active(&self, site: &Arc<Site>)

fn site(&self, id: &DeploymentHash) -> Result<Arc<Site>, StoreError>

Return the active Site for this deployment hash

fn find_site(&self, id: DeploymentId) -> Result<Arc<Site>, StoreError>

fn store(
&self,
id: &DeploymentHash
) -> Result<(&Arc<DeploymentStore>, Arc<Site>), StoreError>

Return the store and site for the active deployment of this deployment hash

fn for_site(&self, site: &Site) -> Result<&Arc<DeploymentStore>, StoreError>

fn layout(&self, id: &DeploymentHash) -> Result<Arc<Layout>, StoreError>

fn place(
&self,
name: &SubgraphName,
network_name: &str,
default_node: NodeId
) -> Result<(Shard, NodeId), StoreError>

fn create_deployment_internal(
&self,
name: SubgraphName,
schema: &Schema,
deployment: SubgraphDeploymentEntity,
node_id: NodeId,
network_name: String,
mode: SubgraphVersionSwitchingMode,
replace: bool
) -> Result<DeploymentLocator, StoreError>

pub fn copy_deployment(
&self,
src: &DeploymentLocator,
shard: Shard,
node: NodeId,
block: BlockPtr
) -> Result<DeploymentLocator, StoreError>

pub fn activate(&self, deployment: &DeploymentLocator) -> Result<(), StoreError>

Mark deployment as the only active deployment amongst all sites with the same deployment hash. Activating this specific deployment will make queries use that instead of whatever was active before

pub fn create_deployment_replace(
&self,
name: SubgraphName,
schema: &Schema,
deployment: SubgraphDeploymentEntity,
node_id: NodeId,
network_name: String,
mode: SubgraphVersionSwitchingMode
) -> Result<DeploymentLocator, StoreError>

pub(crate) fn send_store_event(
&self,
event: &StoreEvent
) -> Result<(), StoreError>

fn primary_conn(&self) -> Result<Connection<'_>, StoreError>

Get a connection to the primary shard. Code must never hold one of these connections while also accessing a DeploymentStore, since both might draw connections from the same pool, and trying to get two connections can deadlock the entire process if the pool runs out of connections in between getting the first one and trying to get the second one.

pub(crate) fn replica_for_query(
&self,
target: QueryTarget,
for_subscription: bool
) -> Result<(Arc<DeploymentStore>, Arc<Site>, ReplicaId), StoreError>

pub fn delete_all_entities_for_test_use_only(&self) -> Result<(), StoreError>

Delete all entities. This function exists solely for integration tests and should never be called from any other code. Unfortunately, Rust makes it very hard to export items just for testing

fn deployments_by_shard(
&self,
sites: Vec<Site>
) -> Result<HashMap<Shard, Vec<Arc<Site>>>, StoreError>

Partition the list of deployments by the shard they belong to. As a side-effect, add all sites to the cache

pub fn record_unused_deployments(
&self
) -> Result<Vec<DeploymentDetail>, StoreError>

Look for new unused deployments and add them to the unused_deployments table

pub fn list_unused_deployments(
&self,
filter: Filter
) -> Result<Vec<UnusedDeployment>, StoreError>

pub fn remove_deployment(&self, id: DeploymentId) -> Result<(), StoreError>

Remove a deployment, i.e., all its data and metadata. This is only permissible if the deployment is unused in the sense that it is neither the current nor pending version of any subgraph, and is not currently assigned to any node

pub(crate) fn status(&self, filter: Filter) -> Result<Vec<Info>, StoreError>

pub(crate) fn version_info(
&self,
version: &str
) -> Result<VersionInfo, StoreError>

pub(crate) fn versions_for_subgraph_id(
&self,
subgraph_id: &str
) -> Result<(Option<String>, Option<String>), StoreError>

pub fn error_count(&self, id: &DeploymentHash) -> Result<usize, StoreError>

pub(crate) async fn vacuum(&self) -> Vec<Result<(), StoreError>>

Vacuum the subgraph_deployment table in each shard

pub fn rewind(
&self,
id: DeploymentHash,
block_ptr_to: BlockPtr
) -> Result<(), StoreError>

pub(crate) fn get_proof_of_indexing<'a>(
self: Arc<Self>,
id: &'a DeploymentHash,
indexer: &'a Option<Address>,
block: BlockPtr
) -> DynTryFuture<'a, Option<[u8; 32]>>

pub fn find(
&self,
query: EntityQuery
) -> Result<Vec<Entity>, QueryExecutionError>

pub fn locate_in_shard(
&self,
hash: &DeploymentHash,
shard: Shard
) -> Result<Option<DeploymentLocator>, StoreError>